requirements.txt
aniso8601==10.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
flask-restx==1.3.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
pymongo==4.12.0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
